internal: Combine build and generate scripts into one index.ts
